Split iOS review notification support tests
