+++ apps/ios/Flashcards/FlashcardsTests/Review/Notifications/ReviewNotificationsSupportTests.swift
@@ -2,594 +2,7 @@ import Foundation
 import XCTest
 @testable import Flashcards
 
-final class ReviewNotificationsSupportTests: XCTestCase {
-    func testDefaultReviewNotificationsSettingsStartEnabled() {
-        let settings = makeDefaultReviewNotificationsSettings()
-
-        XCTAssertTrue(settings.isEnabled)
-        XCTAssertEqual(settings.selectedMode, .daily)
-        XCTAssertEqual(settings.daily.hour, defaultDailyReminderHour)
-        XCTAssertEqual(settings.daily.minute, defaultDailyReminderMinute)
-    }
-
-    func testLoadReviewNotificationsSettingsDefaultsToEnabledWhenUnset() {
-        let suiteName = "ReviewNotificationsSupportTests-\(UUID().uuidString)"
-        guard let userDefaults = UserDefaults(suiteName: suiteName) else {
-            XCTFail("Expected isolated UserDefaults suite")
-            return
-        }
-        defer {
-            userDefaults.removePersistentDomain(forName: suiteName)
-        }
-
-        let settings = loadReviewNotificationsSettings(
-            userDefaults: userDefaults,
-            encoder: JSONEncoder(),
-            decoder: JSONDecoder(),
-            workspaceId: "workspace-1"
-        )
-
-        XCTAssertTrue(settings.isEnabled)
-        XCTAssertEqual(settings.selectedMode, .daily)
-    }
-
-    func testLoadReviewNotificationsSettingsMigratesCurrentWorkspaceOnce() throws {
-        let suiteName = "ReviewNotificationsSupportTests-\(UUID().uuidString)"
-        guard let userDefaults = UserDefaults(suiteName: suiteName) else {
-            XCTFail("Expected isolated UserDefaults suite")
-            return
-        }
-        defer {
-            userDefaults.removePersistentDomain(forName: suiteName)
-        }
-
-        let workspaceId = "workspace-1"
-        let disabledSettings = ReviewNotificationsSettings(
-            isEnabled: false,
-            selectedMode: .inactivity,
-            daily: DailyReviewNotificationsSettings(
-                hour: defaultDailyReminderHour,
-                minute: defaultDailyReminderMinute
-            ),
-            inactivity: InactivityReviewNotificationsSettings(
-                windowStartHour: 9,
-                windowStartMinute: 0,
-                windowEndHour: 19,
-                windowEndMinute: 0,
-                idleMinutes: 60
-            ),
-            showAppIconBadge: true
-        )
-
-        let persistedData = try JSONEncoder().encode(disabledSettings)
-        userDefaults.set(
-            persistedData,
-            forKey: makeLegacyReviewNotificationsSettingsUserDefaultsKey(workspaceId: workspaceId)
-        )
-
-        let loadedSettings = loadReviewNotificationsSettings(
-            userDefaults: userDefaults,
-            encoder: JSONEncoder(),
-            decoder: JSONDecoder(),
-            workspaceId: workspaceId
-        )
-
-        XCTAssertEqual(loadedSettings, disabledSettings)
-        XCTAssertNotNil(userDefaults.data(forKey: reviewNotificationsSettingsUserDefaultsKey))
-
-        let settingsAfterWorkspaceSwitch = loadReviewNotificationsSettings(
-            userDefaults: userDefaults,
-            encoder: JSONEncoder(),
-            decoder: JSONDecoder(),
-            workspaceId: "workspace-2"
-        )
-
-        XCTAssertEqual(settingsAfterWorkspaceSwitch, disabledSettings)
-    }
-
-    func testDefaultStrictRemindersSettingsStartEnabled() {
-        let settings = makeDefaultStrictRemindersSettings()
-
-        XCTAssertTrue(settings.isEnabled)
-    }
-
-    func testReviewNotificationPendingLimitReservesStrictReminderCapacityWhenEnabled() {
-        let settings = StrictRemindersSettings(isEnabled: true)
-
-        XCTAssertEqual(
-            reviewNotificationPendingRequestsLimit(strictRemindersSettings: settings),
-            appNotificationPendingRequestsLimit - strictReminderPendingRequestsLimit
-        )
-    }
-
-    func testReviewNotificationPendingLimitUsesFullLimitWhenStrictRemindersDisabled() {
-        let settings = StrictRemindersSettings(isEnabled: false)
-
-        XCTAssertEqual(
-            reviewNotificationPendingRequestsLimit(strictRemindersSettings: settings),
-            appNotificationPendingRequestsLimit
-        )
-    }
-
-    func testStrictReminderNotificationScopePersistsAndValidatesCurrentNotification() {
-        let suiteName = "ReviewNotificationsSupportTests-\(UUID().uuidString)"
-        guard let userDefaults = UserDefaults(suiteName: suiteName) else {
-            XCTFail("Expected isolated UserDefaults suite")
-            return
-        }
-        defer {
-            userDefaults.removePersistentDomain(forName: suiteName)
-        }
-
-        let scope = loadStrictReminderNotificationScope(userDefaults: userDefaults)
-        let userInfo = buildStrictReminderNotificationUserInfo(scope: scope)
-
-        XCTAssertFalse(scope.isEmpty)
-        XCTAssertEqual(loadStrictReminderNotificationScope(userDefaults: userDefaults), scope)
-        XCTAssertTrue(isCurrentStrictReminderNotification(userInfo: userInfo, userDefaults: userDefaults))
-    }
-
-    func testStrictReminderNotificationScopeRejectsMissingOrRotatedScope() {
-        let suiteName = "ReviewNotificationsSupportTests-\(UUID().uuidString)"
-        guard let userDefaults = UserDefaults(suiteName: suiteName) else {
-            XCTFail("Expected isolated UserDefaults suite")
-            return
-        }
-        defer {
-            userDefaults.removePersistentDomain(forName: suiteName)
-        }
-
-        let initialScope = loadStrictReminderNotificationScope(userDefaults: userDefaults)
-        let initialUserInfo = buildStrictReminderNotificationUserInfo(scope: initialScope)
-
-        XCTAssertFalse(
-            isCurrentStrictReminderNotification(
-                userInfo: [appNotificationTapTypeUserInfoKey: AppNotificationTapType.strictReminder.rawValue],
-                userDefaults: userDefaults
-            )
-        )
-
-        rotateStrictReminderNotificationScope(userDefaults: userDefaults)
-
-        XCTAssertFalse(isCurrentStrictReminderNotification(userInfo: initialUserInfo, userDefaults: userDefaults))
-    }
-
-    func testShouldRemoveStrictReminderNotificationMatchesOnlyLegacyOrCapturedScope() {
-        let oldScope = "old-scope"
-        let newScope = "new-scope"
-
-        XCTAssertTrue(
-            shouldRemoveStrictReminderNotification(
-                userInfo: buildStrictReminderNotificationUserInfo(scope: oldScope),
-                removalScope: oldScope
-            )
-        )
-        XCTAssertTrue(
-            shouldRemoveStrictReminderNotification(
-                userInfo: [appNotificationTapTypeUserInfoKey: AppNotificationTapType.strictReminder.rawValue],
-                removalScope: oldScope
-            )
-        )
-        XCTAssertFalse(
-            shouldRemoveStrictReminderNotification(
-                userInfo: buildStrictReminderNotificationUserInfo(scope: newScope),
-                removalScope: oldScope
-            )
-        )
-        XCTAssertFalse(
-            shouldRemoveStrictReminderNotification(
-                userInfo: [appNotificationTapTypeUserInfoKey: AppNotificationTapType.reviewReminder.rawValue],
-                removalScope: oldScope
-            )
-        )
-    }
-
-    func testStrictReminderRemovalScopesRemoveCurrentScopeAndLegacyPayloads() {
-        XCTAssertEqual(
-            strictReminderRemovalScopes(currentScope: "current-scope"),
-            ["current-scope", nil]
-        )
-        XCTAssertEqual(
-            strictReminderRemovalScopes(currentScope: nil),
-            [nil]
-        )
-        XCTAssertEqual(
-            strictReminderRemovalScopes(currentScope: ""),
-            [nil]
-        )
-    }
-
-    func testMakeStrictRemindersReconcileRequestCarriesClearDeliveredFlagFromTrigger() throws {
-        let calendar = makeCalendar()
-        let now = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 3, hour: 21, minute: 5, calendar: calendar))
-
-        XCTAssertEqual(
-            makeStrictRemindersReconcileRequest(
-                trigger: .appActive,
-                now: now,
-                shouldClearDeliveredStrictReminders: false
-            ),
-            StrictRemindersReconcileRequest(
-                now: now,
-                triggers: [.appActive],
-                shouldClearDeliveredStrictReminders: true
-            )
-        )
-        XCTAssertEqual(
-            makeStrictRemindersReconcileRequest(
-                trigger: .reviewRecorded,
-                now: now,
-                shouldClearDeliveredStrictReminders: false
-            ),
-            StrictRemindersReconcileRequest(
-                now: now,
-                triggers: [.reviewRecorded],
-                shouldClearDeliveredStrictReminders: false
-            )
-        )
-        XCTAssertEqual(
-            makeStrictRemindersReconcileRequest(
-                trigger: .workspaceChanged,
-                now: now,
-                shouldClearDeliveredStrictReminders: false
-            ),
-            StrictRemindersReconcileRequest(
-                now: now,
-                triggers: [.workspaceChanged],
-                shouldClearDeliveredStrictReminders: true
-            )
-        )
-        XCTAssertTrue(
-            makeStrictRemindersReconcileRequest(
-                trigger: .settingsChanged,
-                now: now,
-                shouldClearDeliveredStrictReminders: true
-            ).shouldClearDeliveredStrictReminders
-        )
-    }
-
-    func testMergeStrictRemindersReconcileRequestsKeepsLatestNowAndPendingDeliveredClear() throws {
-        let calendar = makeCalendar()
-        let earlierNow = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 3, hour: 21, minute: 5, calendar: calendar))
-        let laterNow = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 3, hour: 21, minute: 6, calendar: calendar))
-
-        let mergedRequest = mergeStrictRemindersReconcileRequests(
-            pendingRequest: StrictRemindersReconcileRequest(
-                now: earlierNow,
-                triggers: [.appActive],
-                shouldClearDeliveredStrictReminders: true
-            ),
-            nextRequest: StrictRemindersReconcileRequest(
-                now: laterNow,
-                triggers: [.reviewRecorded],
-                shouldClearDeliveredStrictReminders: false
-            )
-        )
-
-        XCTAssertEqual(
-            mergedRequest,
-            StrictRemindersReconcileRequest(
-                now: laterNow,
-                triggers: [.appActive, .reviewRecorded],
-                shouldClearDeliveredStrictReminders: true
-            )
-        )
-        XCTAssertEqual(
-            strictRemindersReconcileTriggerDiagnosticValue(triggers: mergedRequest.triggers),
-            "app_active+review_recorded"
-        )
-    }
-
-    func testResolveStrictReminderCompletedDayResolutionIncludesImportedCurrentDayReview() async throws {
-        let suiteName = "ReviewNotificationsSupportTests-\(UUID().uuidString)"
-        guard let userDefaults = UserDefaults(suiteName: suiteName) else {
-            XCTFail("Expected isolated UserDefaults suite")
-            return
-        }
-        let (database, databaseURL) = try makeTemporaryLocalDatabase()
-        defer {
-            userDefaults.removePersistentDomain(forName: suiteName)
-            try? database.close()
-            try? removeTemporaryDatabase(at: databaseURL)
-        }
-
-        let workspace = try database.workspaceSettingsStore.loadWorkspace()
-        let card = try database.saveCard(
-            workspaceId: workspace.workspaceId,
-            input: CardEditorInput(
-                frontText: "Question",
-                backText: "Answer",
-                tags: [],
-            ),
-            cardId: nil,
-            mediaAssetIdsReadyForUpload: []
-        )
-        let calendar = makeCalendar()
-        let reviewedAt = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 3, hour: 10, minute: 15, calendar: calendar))
-        let now = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 3, hour: 21, minute: 5, calendar: calendar))
-
-        _ = try database.submitReview(
-            workspaceId: workspace.workspaceId,
-            reviewSubmission: ReviewSubmission(
-                cardId: card.cardId,
-                rating: .good,
-                reviewedAtClient: formatIsoTimestamp(date: reviewedAt),
-                reviewedTimeZone: "UTC"
-            )
-        )
-
-        let resolution = resolveStrictReminderCompletedDayResolution(
-            persistedCompletedDayStartMillis: loadStrictReminderCompletedDayStartMillis(
-                userDefaults: userDefaults,
-                now: now,
-                calendar: calendar
-            ),
-            importedCompletedDayStartMillis: try await loadStrictReminderImportedCompletedDayStartMillis(
-                databaseURL: databaseURL,
-                workspaceId: workspace.workspaceId,
-                now: now,
-                calendar: calendar
-            ),
-            prefersImportedCurrentDayCompletion: true
-        )
-
-        XCTAssertEqual(
-            resolution.completedDayStartMillis,
-            [strictReminderDayStartMillis(date: calendar.startOfDay(for: now))]
-        )
-        XCTAssertTrue(resolution.shouldPersistImportedCompletion)
-        XCTAssertFalse(resolution.shouldClearPersistedCompletion)
-        XCTAssertEqual(
-            try await loadStrictReminderImportedCompletedDayStartMillis(
-                databaseURL: databaseURL,
-                workspaceId: "workspace-without-review",
-                now: now,
-                calendar: calendar
-            ),
-            []
-        )
-    }
-
-    func testResolveStrictReminderCompletedDayResolutionDoesNotRePersistExistingCurrentDayCompletion() async throws {
-        let suiteName = "ReviewNotificationsSupportTests-\(UUID().uuidString)"
-        guard let userDefaults = UserDefaults(suiteName: suiteName) else {
-            XCTFail("Expected isolated UserDefaults suite")
-            return
-        }
-        let (database, databaseURL) = try makeTemporaryLocalDatabase()
-        defer {
-            userDefaults.removePersistentDomain(forName: suiteName)
-            try? database.close()
-            try? removeTemporaryDatabase(at: databaseURL)
-        }
-
-        let workspace = try database.workspaceSettingsStore.loadWorkspace()
-        let card = try database.saveCard(
-            workspaceId: workspace.workspaceId,
-            input: CardEditorInput(
-                frontText: "Question",
-                backText: "Answer",
-                tags: [],
-            ),
-            cardId: nil,
-            mediaAssetIdsReadyForUpload: []
-        )
-        let calendar = makeCalendar()
-        let reviewedAt = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 3, hour: 10, minute: 15, calendar: calendar))
-        let now = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 3, hour: 21, minute: 5, calendar: calendar))
-
-        persistStrictReminderLastReviewedAt(
-            userDefaults: userDefaults,
-            reviewedAt: reviewedAt
-        )
-        _ = try database.submitReview(
-            workspaceId: workspace.workspaceId,
-            reviewSubmission: ReviewSubmission(
-                cardId: card.cardId,
-                rating: .good,
-                reviewedAtClient: formatIsoTimestamp(date: reviewedAt),
-                reviewedTimeZone: "UTC"
-            )
-        )
-
-        let resolution = resolveStrictReminderCompletedDayResolution(
-            persistedCompletedDayStartMillis: loadStrictReminderCompletedDayStartMillis(
-                userDefaults: userDefaults,
-                now: now,
-                calendar: calendar
-            ),
-            importedCompletedDayStartMillis: try await loadStrictReminderImportedCompletedDayStartMillis(
-                databaseURL: databaseURL,
-                workspaceId: workspace.workspaceId,
-                now: now,
-                calendar: calendar
-            ),
-            prefersImportedCurrentDayCompletion: true
-        )
-
-        XCTAssertEqual(
-            resolution.completedDayStartMillis,
-            [strictReminderDayStartMillis(date: calendar.startOfDay(for: now))]
-        )
-        XCTAssertFalse(resolution.shouldPersistImportedCompletion)
-        XCTAssertFalse(resolution.shouldClearPersistedCompletion)
-    }
-
-    func testResolveStrictReminderCompletedDayResolutionClearsStalePersistedCurrentDayCompletionWhenDatabaseHasNoReviewRows() async throws {
-        let suiteName = "ReviewNotificationsSupportTests-\(UUID().uuidString)"
-        guard let userDefaults = UserDefaults(suiteName: suiteName) else {
-            XCTFail("Expected isolated UserDefaults suite")
-            return
-        }
-        let (database, databaseURL) = try makeTemporaryLocalDatabase()
-        defer {
-            userDefaults.removePersistentDomain(forName: suiteName)
-            try? database.close()
-            try? removeTemporaryDatabase(at: databaseURL)
-        }
-
-        let workspace = try database.workspaceSettingsStore.loadWorkspace()
-        let card = try database.saveCard(
-            workspaceId: workspace.workspaceId,
-            input: CardEditorInput(
-                frontText: "Question",
-                backText: "Answer",
-                tags: [],
-            ),
-            cardId: nil,
-            mediaAssetIdsReadyForUpload: []
-        )
-        let calendar = makeCalendar()
-        let reviewedAt = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 3, hour: 10, minute: 15, calendar: calendar))
-        let now = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 3, hour: 21, minute: 5, calendar: calendar))
-
-        persistStrictReminderLastReviewedAt(
-            userDefaults: userDefaults,
-            reviewedAt: reviewedAt
-        )
-        _ = try database.submitReview(
-            workspaceId: workspace.workspaceId,
-            reviewSubmission: ReviewSubmission(
-                cardId: card.cardId,
-                rating: .good,
-                reviewedAtClient: formatIsoTimestamp(date: reviewedAt),
-                reviewedTimeZone: "UTC"
-            )
-        )
-        _ = try database.core.execute(
-            sql: "DELETE FROM review_events WHERE workspace_id = ?",
-            values: [.text(workspace.workspaceId)]
-        )
-
-        let resolution = resolveStrictReminderCompletedDayResolution(
-            persistedCompletedDayStartMillis: loadStrictReminderCompletedDayStartMillis(
-                userDefaults: userDefaults,
-                now: now,
-                calendar: calendar
-            ),
-            importedCompletedDayStartMillis: try await loadStrictReminderImportedCompletedDayStartMillis(
-                databaseURL: databaseURL,
-                workspaceId: workspace.workspaceId,
-                now: now,
-                calendar: calendar
-            ),
-            prefersImportedCurrentDayCompletion: true
-        )
-
-        XCTAssertEqual(resolution.completedDayStartMillis, [])
-        XCTAssertFalse(resolution.shouldPersistImportedCompletion)
-        XCTAssertTrue(resolution.shouldClearPersistedCompletion)
-    }
-
-    func testLoadStrictReminderCompletedDayStartMillisUsesPersistedReviewWithinCurrentLocalDay() throws {
-        let suiteName = "ReviewNotificationsSupportTests-\(UUID().uuidString)"
-        guard let userDefaults = UserDefaults(suiteName: suiteName) else {
-            XCTFail("Expected isolated UserDefaults suite")
-            return
-        }
-        defer {
-            userDefaults.removePersistentDomain(forName: suiteName)
-        }
-
-        let calendar = makeCalendar()
-        let reviewedAt = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 3, hour: 10, minute: 15, calendar: calendar))
-        let now = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 3, hour: 21, minute: 5, calendar: calendar))
-        let expectedDayStart = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 3, hour: 0, minute: 0, calendar: calendar))
-
-        persistStrictReminderLastReviewedAt(
-            userDefaults: userDefaults,
-            reviewedAt: reviewedAt
-        )
-
-        XCTAssertEqual(
-            loadStrictReminderCompletedDayStartMillis(
-                userDefaults: userDefaults,
-                now: now,
-                calendar: calendar
-            ),
-            [strictReminderDayStartMillis(date: expectedDayStart)]
-        )
-    }
-
-    func testLoadStrictReminderCompletedDayStartMillisIgnoresPersistedReviewOutsideCurrentLocalDay() throws {
-        let suiteName = "ReviewNotificationsSupportTests-\(UUID().uuidString)"
-        guard let userDefaults = UserDefaults(suiteName: suiteName) else {
-            XCTFail("Expected isolated UserDefaults suite")
-            return
-        }
-        defer {
-            userDefaults.removePersistentDomain(forName: suiteName)
-        }
-
-        let calendar = makeCalendar()
-        let reviewedAt = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 2, hour: 23, minute: 30, calendar: calendar))
-        let now = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 3, hour: 21, minute: 5, calendar: calendar))
-
-        persistStrictReminderLastReviewedAt(
-            userDefaults: userDefaults,
-            reviewedAt: reviewedAt
-        )
-
-        XCTAssertEqual(
-            loadStrictReminderCompletedDayStartMillis(
-                userDefaults: userDefaults,
-                now: now,
-                calendar: calendar
-            ),
-            []
-        )
-    }
-
-    func testClearStoredStrictRemindersRemovesCompletionAndScheduledPayloads() throws {
-        let suiteName = "ReviewNotificationsSupportTests-\(UUID().uuidString)"
-        guard let userDefaults = UserDefaults(suiteName: suiteName) else {
-            XCTFail("Expected isolated UserDefaults suite")
-            return
-        }
-        defer {
-            userDefaults.removePersistentDomain(forName: suiteName)
-        }
-
-        let calendar = makeCalendar()
-        let reviewedAt = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 3, hour: 10, minute: 15, calendar: calendar))
-        let now = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 3, hour: 21, minute: 5, calendar: calendar))
-        let payload = ScheduledStrictReminderPayload(
-            dayStartMillis: strictReminderDayStartMillis(
-                date: calendar.startOfDay(for: now)
-            ),
-            scheduledAtMillis: Int64(now.timeIntervalSince1970 * 1_000),
-            offset: .twoHours,
-            requestId: "strict-reminder::2h::2026-04-03-22-00"
-        )
-
-        persistStrictReminderLastReviewedAt(
-            userDefaults: userDefaults,
-            reviewedAt: reviewedAt
-        )
-        let payloadData = try JSONEncoder().encode([payload])
-        userDefaults.set(payloadData, forKey: strictReminderScheduledPayloadsUserDefaultsKey)
-
-        clearStoredStrictReminders(userDefaults: userDefaults)
-
-        XCTAssertEqual(
-            loadStrictReminderCompletedDayStartMillis(
-                userDefaults: userDefaults,
-                now: now,
-                calendar: calendar
-            ),
-            []
-        )
-        XCTAssertEqual(
-            loadScheduledStrictReminders(
-                userDefaults: userDefaults,
-                decoder: JSONDecoder()
-            ),
-            []
-        )
-        XCTAssertNil(userDefaults.object(forKey: strictReminderLastReviewedAtUserDefaultsKey))
-        XCTAssertNil(userDefaults.object(forKey: strictReminderScheduledPayloadsUserDefaultsKey))
-    }
-
+final class ReviewNotificationsSupportTests: ReviewNotificationsTestCase {
     func testInactivityReminderDatesRepeatAcrossCurrentAndLaterDays() throws {
         let calendar = makeCalendar()
         let lastActiveAt = try XCTUnwrap(makeDate(year: 2026, month: 4, day: 3, hour: 10, minute: 15, calendar: calendar))
@@ -1130,61 +543,4 @@ final class ReviewNotificationsSupportTests: XCTestCase {
             .owned(.openStrictReminder)
         )
     }
-}
-
-private func makeCalendar() -> Calendar {
-    var calendar = Calendar(identifier: .gregorian)
-    calendar.timeZone = TimeZone(secondsFromGMT: 0) ?? .gmt
-    calendar.locale = Locale(identifier: "en_US_POSIX")
-    return calendar
-}
-
-private func makeDate(
-    year: Int,
-    month: Int,
-    day: Int,
-    hour: Int,
-    minute: Int,
-    calendar: Calendar
-) -> Date? {
-    calendar.date(
-        from: DateComponents(
-            calendar: calendar,
-            timeZone: calendar.timeZone,
-            year: year,
-            month: month,
-            day: day,
-            hour: hour,
-            minute: minute
-        )
-    )
-}
-
-private func formatDate(date: Date, calendar: Calendar) -> String {
-    let formatter = DateFormatter()
-    formatter.calendar = calendar
-    formatter.locale = Locale(identifier: "en_US_POSIX")
-    formatter.timeZone = calendar.timeZone
-    formatter.dateFormat = "yyyy-MM-dd HH:mm"
-    return formatter.string(from: date)
-}
-
-private func makeTemporaryLocalDatabase() throws -> (database: LocalDatabase, databaseURL: URL) {
-    let databaseDirectory = FileManager.default.temporaryDirectory
-        .appendingPathComponent(UUID().uuidString.lowercased(), isDirectory: true)
-    try FileManager.default.createDirectory(
-        at: databaseDirectory,
-        withIntermediateDirectories: true,
-        attributes: nil
-    )
-    let databaseURL = databaseDirectory.appendingPathComponent("flashcards.sqlite", isDirectory: false)
-    return (try LocalDatabase(databaseURL: databaseURL), databaseURL)
-}
-
-private func removeTemporaryDatabase(at databaseURL: URL) throws {
-    try FileManager.default.removeItem(at: databaseURL.deletingLastPathComponent())
-}
-
-private extension PersistedReviewFilter {
-    static let allCards = PersistedReviewFilter(kind: .allCards, deckId: nil, effortLevel: nil, tag: nil)
 }
